+++ design-pattern/src/main/java/com/imugen/design/behavioral/strategy/after/Strategy.java
@@ -4,7 +4,7 @@ package com.imugen.design.behavioral.strategy.after;
  * @author Daniel Liu 2020/2/25 10:24
  */
 public interface Strategy {
-    double compute(long money);
+    long compute(long money);
 
     int getType();
 }
